feat(drive): same block execution model
