+++ packages/dashmate/configs/system/base.js
@@ -131,7 +131,7 @@ module.exports = {
       },
       tenderdash: {
         docker: {
-          image: 'kostya/tenderdash',
+          image: 'dashpay/tenderdash:0.10.0-dev.2',
         },
         p2p: {
           port: 26656,
