+++ packages/js-drive/lib/abci/handlers/finalizeBlockHandlerFactory.js
@@ -5,20 +5,27 @@ const {
     },
   },
 } = require('@dashevo/abci/types');
+const ReadOperation = require('@dashevo/dpp/lib/stateTransition/fee/operations/ReadOperation');
+const BlockExecutionContext = require('../../blockExecution/BlockExecutionContext');
+const DataContractCacheItem = require('../../dataContract/DataContractCacheItem');
 
 /**
  *
  * @return {finalizeBlockHandler}
  * @param {GroveDBStore} groveDBStore
  * @param {BlockExecutionContext} blockExecutionContext
- * @param {commit} commit
+ * @param {BlockExecutionContextRepository} blockExecutionContextRepository
+ * @param {LRUCache} dataContractCache
+ * @param {CoreRpcClient} coreRpcClient
  * @param {BaseLogger} logger
  * @param {ExecutionTimer} executionTimer
  */
 function finalizeBlockHandlerFactory(
   groveDBStore,
   blockExecutionContext,
-  commit,
+  blockExecutionContextRepository,
+  dataContractCache,
+  coreRpcClient,
   logger,
   executionTimer,
 ) {
@@ -29,11 +36,6 @@ function finalizeBlockHandlerFactory(
    * @return {Promise<abci.ResponseFinalizeBlock>}
    */
   async function finalizeBlockHandler(request) {
-    // Start block execution timer
-    executionTimer.clearTimer('blockExecution');
-
-    executionTimer.startTimer('blockExecution');
-
     const {
       decidedLastCommit: lastCommitInfo,
       height,
@@ -53,7 +55,56 @@ function finalizeBlockHandlerFactory(
     blockExecutionContext.setPreviousHeight(height);
     blockExecutionContext.setPreviousCoreChainLockedHeight(coreChainLockedHeight);
 
-    await commit(lastCommitInfo, consensusLogger);
+    consensusLogger.debug('Commit ABCI method requested');
+
+    // Store block execution context
+    const clonedBlockExecutionContext = new BlockExecutionContext();
+    clonedBlockExecutionContext.populate(blockExecutionContext);
+
+    blockExecutionContextRepository.store(
+      clonedBlockExecutionContext,
+      {
+        useTransaction: true,
+      },
+    );
+
+    // Commit the current block db transactions
+    await groveDBStore.commitTransaction();
+
+    // Update data contract cache with new version of
+    // committed data contract
+    for (const dataContract of blockExecutionContext.getDataContracts()) {
+      const operations = [new ReadOperation(dataContract.toBuffer().length)];
+
+      const cacheItem = new DataContractCacheItem(dataContract, operations);
+
+      if (dataContractCache.has(cacheItem.getKey())) {
+        dataContractCache.set(cacheItem.getKey(), cacheItem);
+      }
+    }
+
+    // Send withdrawal transactions to Core
+    const unsignedWithdrawalTransactionsMap = blockExecutionContext.getWithdrawalTransactionsMap();
+
+    const { thresholdVoteExtensions } = lastCommitInfo;
+
+    for (const { extension, signature } of (thresholdVoteExtensions || [])) {
+      const withdrawalTransactionHash = extension.toString('hex');
+
+      const unsignedWithdrawalTransactionBytes = unsignedWithdrawalTransactionsMap[
+        withdrawalTransactionHash
+      ];
+
+      if (unsignedWithdrawalTransactionBytes) {
+        const transactionBytes = Buffer.concat([
+          unsignedWithdrawalTransactionBytes,
+          signature,
+        ]);
+
+        // TODO: think about Core error handling
+        await coreRpcClient.sendRawTransaction(transactionBytes.toString('hex'));
+      }
+    }
 
     const blockExecutionTimings = executionTimer.stopTimer('blockExecution');
     const blockHeight = blockExecutionContext.getHeight();
